detection_2/requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
requests
pillow
pydantic